requirements.txt
aiogram==3.4.1
aiohttp==3.9.1
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-dotenv==1.0.0
groq==0.4.1
openai==1.10.0
pydantic==2.5.3
asyncio==3.4.3
